# Data Dictinary Open API BRD files/openapi.yaml
openapi: 3.0.3
info:
  title: IAMS API
  description: >
    Inventory Audit Management System REST API. Traces to IAMS-FRS-3.0 (functional
    requirements), IAMS-SRS-3.0 (architecture/security), and IAMS-DD-1.0 (schema).
    All endpoints are served on-premises; no endpoint has a mandatory external
    dependency. Authorization is enforced server-side on every endpoint per
    SRS Section 6.2 regardless of what the client UI hides.
  version: "1.0.0"
  contact:
    name: IAMS Engineering
servers:
  - url: /api/v1
    description: On-premises deployment, versioned base path
security:
  - bearerAuth: []
...
paths:
...
  /assets:
    get:
      tags: [Assets]
      summary: List assets (paginated, filterable) — FR-RPT-01 filters, NFR-SCALE-03
      parameters:
        - { name: category_id, in: query, schema: { type: string, format: uuid } }
        - { name: department_id, in: query, schema: { type: string, format: uuid } }
        - { name: location_node_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        '200':
          description: Paginated asset list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Asset' } }
                  total: { type: integer }
                  page: { type: integer }
    post:
      tags: [Assets]
      summary: Register a new asset — FR-AST-01, 02, 05, 06; see UC-AST-01
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetCreateRequest' }
      responses:
        '201':
          description: Asset created, QR code generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '400': { description: Validation error — see Error schema }
  /assets/{assetId}:
    parameters:
      - { name: assetId, in: path, required: true, schema: { type: string } }
    get:
      tags: [Assets]
      summary: Get asset detail, including history and attachments — FR-AST-08 through 15
      responses:
        '200':
          description: Asset detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetDetail' }
        '404': { description: Not found }
    put:
      tags: [Assets]
      summary: Update asset (requires If-Match version header — NFR-CONC-01)
      parameters:
        - { name: If-Match, in: header, required: true, schema: { type: string }, description: "Current version, for optimistic locking" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetUpdateRequest' }
      responses:
        '200': { description: Updated }
        '409': { description: Version conflict — asset was modified concurrently }
  /assets/{assetId}/history:
    get:
      tags: [Assets]
      summary: Append-only change history — FR-AST-10
      parameters: [{ name: assetId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: History entries, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AssetHistoryEntry' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FR-SEC-01. Service accounts for integrations use scoped credentials per SRS Section 6.6, never a shared administrative token.
  schemas:
...
    Asset:
      type: object
      properties:
        asset_id: { type: string, format: uuid }
        asset_number: { type: string, example: "AST-00482" }
        name: { type: string }
        category_id: { type: string, format: uuid }
        status: { type: string, enum: [In Use, In Storage, Under Repair, Missing, Retired, Disposed] }
        location_node_id: { type: string, format: uuid }
        department_id: { type: string, format: uuid }
        assigned_to_user_id: { type: string, format: uuid, nullable: true }
        qr_code_value: { type: string }
        version: { type: integer, description: "Optimistic locking — NFR-CONC-01" }
    AssetDetail:
      allOf:
        - $ref: '#/components/schemas/Asset'
        - type: object
          properties:
            manufacturer_id: { type: string, format: uuid, nullable: true }
            vendor_id: { type: string, format: uuid, nullable: true }
            purchase_cost: { type: number }
            currency: { type: string, example: USD }
            purchase_date: { type: string, format: date }
            warranty_expiry: { type: string, format: date, nullable: true }
            amc_expiry: { type: string, format: date, nullable: true }
            custom_fields: { type: object }
            insurance_policy: { $ref: '#/components/schemas/InsurancePolicy' }
            attachments: { type: array, items: { type: object, properties: { attachment_id: { type: string }, object_store_key: { type: string } } } }
    AssetCreateRequest:
      type: object
      required: [name, category_id, location_node_id, department_id]
      properties:
        name: { type: string }
        category_id: { type: string, format: uuid }
        manufacturer_id: { type: string, format: uuid }
        vendor_id: { type: string, format: uuid }
        location_node_id: { type: string, format: uuid }
        department_id: { type: string, format: uuid }
        purchase_cost: { type: number }
        currency: { type: string }
        purchase_date: { type: string, format: date }
        custom_fields: { type: object }
    AssetUpdateRequest:
      type: object
      properties:
        name: { type: string }
        status: { type: string }
        custom_fields: { type: object }
    AssetHistoryEntry:
      type: object
      properties:
        history_id: { type: string, format: uuid }
        event_type: { type: string }
        field_changed: { type: string, nullable: true }
        old_value: { type: string, nullable: true }
        new_value: { type: string, nullable: true }
        changed_by: { type: string, format: uuid }
        changed_at: { type: string, format: date-time }
    InsurancePolicy:
      type: object
      properties:
        insurer: { type: string }
        policy_number: { type: string }
        coverage_amount: { type: number }
        expiry: { type: string, format: date }
...